// src/components/landing-page/components/block/block.enum.ts
enum BLOCK {
  ACCORDION = '@@@Accordion@@@',
  ADVANTAGE_LIST = '@@@AdvantageList@@@',
  ADVANTAGE_LIST_NO_BUTTON = '@@@AdvantageListNoButton@@@',
  CONTACT_US = '@@@ContactUs@@@',
  BUTTON = '@@@Button@@@',
  HOW_TO_GO_NEXT = '@@@HowToGoNext@@@',
  LOGO_CLOUD = '@@@LogoCloud@@@',
  PLAYER = '@@@Player@@@',
  REVIEW = '@@@Review@@@',
  ORDERED_LIST = '@@@OrderedList@@@',
  LOGOS = '@@@Logos@@@',
  QUESTION = '@@@Question@@@',
}

export default BLOCK;